requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pillow
python-multipart
